requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.3.1
django-countries==7.5.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
idna==3.6
phonenumbers==8.13.27
pillow==10.2.0
python-decouple==3.8
python-magic==0.4.27
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
